issue-84: Fixed Score Rendering - Added longToStr function to handle casting - Changed score data types from int to long to handle 6 digit numbers

+++ obj Logic/score.h
@@ -5,8 +5,8 @@ typedef struct { /* Structure for the Score Counter */
 
     int visible; /* No = 0, Yes = 1 */
     int posX, posY; /* Current Position */
-    int value; /* Current Score */
-    int highValue; /* High Score */
+    long value; /* Current Score */
+    long highValue; /* High Score */
 
 } Score;
